lineshape out of ans.Data (as they are not data) moving enum definitions into separate files as they are becomning unwieldy
