--- archnemesis/Data/enum/__init__.py
@@ -1,3 +1,0 @@
-"""
-Module contains ENUMs that are used to select/record states such as which lineshape function to use.
-"""
