+++ archnemesis/enum/__init__.py
@@ -1,0 +1,27 @@
+"""
+Module contains ENUMs that are used to select/record states such as which lineshape function to use.
+"""
+
+from .aerosol_phase_function_calculation_mode_enum import AerosolPhaseFunctionCalculationModeEnum # noqa: F401 :: Unused variable is imported to namespace
+from .ambient_gas_enum                             import AmbientGasEnum                          # noqa: F401 :: Unused variable is imported to namespace
+from .atmospheric_profile_format_enum              import AtmosphericProfileFormatEnum            # noqa: F401 :: Unused variable is imported to namespace
+from .atmospheric_profile_type_enum                import AtmosphericProfileTypeEnum              # noqa: F401 :: Unused variable is imported to namespace
+from .emission_type_enum                           import EmissionTypeEnum                        # noqa: F401 :: Unused variable is imported to namespace
+from .gas_enum                                     import GasEnum                                 # noqa: F401 :: Unused variable is imported to namespace
+from .instrument_lineshape_enum                    import InstrumentLineshapeEnum                 # noqa: F401 :: Unused variable is imported to namespace
+from .interpolation_method_enum                    import InterpolationMethodEnum                 # noqa: F401 :: Unused variable is imported to namespace
+from .layer_integration_scheme_enum                import LayerIntegrationSchemeEnum              # noqa: F401 :: Unused variable is imported to namespace
+from .layer_type_enum                              import LayerTypeEnum                           # noqa: F401 :: Unused variable is imported to namespace
+from .lower_boundary_condition_enum                import LowerBoundaryConditionEnum              # noqa: F401 :: Unused variable is imported to namespace
+from .para_H2_ratio_enum                           import ParaH2RatioEnum                         # noqa: F401 :: Unused variable is imported to namespace
+from .path_calc_enum                               import PathCalcEnum                            # noqa: F401 :: Unused variable is imported to namespace
+from .path_observer_pointing_enum                  import PathObserverPointingEnum                # noqa: F401 :: Unused variable is imported to namespace
+from .planet_enum                                  import PlanetEnum                              # noqa: F401 :: Unused variable is imported to namespace
+from .rayleigh_scattering_mode_enum                import RayleighScatteringModeEnum              # noqa: F401 :: Unused variable is imported to namespace
+from .retrieval_strategy_enum                      import RetrievalStrategyEnum                   # noqa: F401 :: Unused variable is imported to namespace
+from .scattering_calculation_mode_enum             import ScatteringCalculationModeEnum           # noqa: F401 :: Unused variable is imported to namespace
+from .spectra_unit_enum                            import SpectraUnitEnum                         # noqa: F401 :: Unused variable is imported to namespace
+from .spectral_calculation_mode_enum               import SpectralCalculationModeEnum             # noqa: F401 :: Unused variable is imported to namespace
+from .spectroscopic_line_profile_enum              import SpectroscopicLineProfileEnum            # noqa: F401 :: Unused variable is imported to namespace
+from .wave_unit_enum                               import WaveUnitEnum                            # noqa: F401 :: Unused variable is imported to namespace
+from .zenith_angle_origin_enum                     import ZenithAngleOriginEnum                   # noqa: F401 :: Unused variable is imported to namespace
