+++ archnemesis/enum/ambient_gas_enum.py
@@ -1,0 +1,11 @@
+from enum import IntEnum
+
+class AmbientGasEnum(IntEnum):
+    """
+    Defines the ambient gas used in the line spectroscopy calculations
+
+    Used as ambient_gas in 'LineData_0.py'
+    """
+    AIR = 0
+    CO2 = 1
+    H2 = 2
