+++ archnemesis/enum/gas_enum.py
@@ -1,0 +1,146 @@
+
+from enum import IntEnum
+
+
+class GasEnum(IntEnum):
+    """
+    Enum to define the gases used in the retrievals.
+    
+    Used in 'IPAIRG1', 'IPAIRG2'
+    """
+    H2O = 1
+    CO2 = 2
+    O3 = 3
+    N2O = 4
+    CO = 5
+    CH4 = 6
+    O2 = 7
+    NO = 8
+    SO2 = 9
+    NO2 = 10
+    NH3 = 11
+    HNO3 = 12
+    OH = 13
+    HF = 14
+    HCl = 15
+    HBr = 16
+    HI = 17
+    ClO = 18
+    OCS = 19
+    H2CO = 20
+    HOCl = 21
+    N2 = 22
+    HCN = 23
+    CH3Cl = 24
+    H2O2 = 25
+    C2H2 = 26
+    C2H6 = 27
+    PH3 = 28
+    C2N2 = 29
+    C4H2 = 30
+    HC3N = 31
+    C2H4 = 32
+    GeH4 = 33
+    C3H8 = 34
+    HCOOH = 35
+    H2S = 36
+    COF2 = 37
+    SF6 = 38
+    H2 = 39
+    He = 40
+    AsH3 = 41
+    C3H4 = 42
+    ClONO2 = 43
+    HO2 = 44
+    O = 45
+    NO_PLUS = 46
+    CH3OH = 47
+    H = 48
+    C6H6 = 49
+    CH3CN = 50
+    CH2NH = 51
+    C2H3CN = 52
+    HCP = 53
+    CS = 54
+    HC5N = 55
+    HC7N = 56
+    C2H5CN = 57
+    CH3NH2 = 58
+    HNC = 59
+    Na = 60
+    K = 61
+    TiO = 62
+    VO = 63
+    CH2CCH2 = 64
+    C4N2 = 65
+    C5H5N = 66
+    C5H4N2 = 67
+    C7H8 = 68
+    C8H6 = 69
+    C5H5CN = 70
+    HOBr = 71
+    CH3Br = 72
+    CF4 = 73
+    SO3 = 74
+    Ne = 75
+    Ar = 76
+    COCl2 = 77
+    SO = 78
+    H2SO4 = 79
+    E_MINUS = 80
+    H3_PLUS = 81
+    FeH = 82
+    AlO = 83
+    AlCl = 84
+    AlF = 85
+    AlH = 86
+    BeH = 87
+    C2 = 88
+    CaF = 89
+    CaH = 90
+    H_MINUS = 91
+    CaO = 92
+    CH = 93
+    CH3 = 94
+    CH3F = 95
+    CN = 96
+    CP = 97
+    CrH = 98
+    HD_PLUS = 99
+    HeH_PLUS = 100
+    KCl = 101
+    KF = 102
+    LiCl = 103
+    LiF = 104
+    LiH = 105
+    LiH_PLUS = 106
+    MgF = 107
+    MgH = 108
+    MgO = 109
+    NaCl = 110
+    NaF = 111
+    NaH = 112
+    NH = 113
+    NS = 114
+    OH_PLUS = 115
+    CIS_P2H2 = 116
+    TRANS_P2H2 = 117
+    PH = 118
+    PN = 119
+    PO = 120
+    PS = 121
+    ScH = 122
+    SH = 123
+    SiH = 124
+    SiH2 = 125
+    SiH4 = 126
+    SiO = 127
+    SiS = 129
+    TiH = 130
+    Cl2 = 131
+    ClO2 = 132
+    BrO = 133
+    IO = 134
+    NO3 = 135
+    BrO2 = 136
+    IO2 = 137
