+++ archnemesis/enum/retrieval_strategy_enum.py
@@ -1,0 +1,8 @@
+from enum import IntEnum
+
+class RetrievalStrategyEnum(IntEnum):
+    """
+    Defines values for 'retrieval_method'
+    """
+    Optimal_Estimation = 0
+    Nested_Sampling = 1
