+++ archnemesis/enum/wave_unit_enum.py
@@ -1,0 +1,8 @@
+from enum import IntEnum
+
+class WaveUnit(IntEnum):
+    """
+    Defines values for 'ISPACE'
+    """
+    Wavenumber_cm = 0
+    Wavelength_um = 1
