+++ archnemesis/lineshape/_const.py
