+++ archnemesis/lineshape/_scipy_support.py
