+++ archnemesis/lineshape/gaussian.py
