+++ archnemesis/lineshape/hartmann_empirical_infrared_CH4_H2_broadening.py
