+++ archnemesis/lineshape/voigt_impl/voigt_scipy.py
